Most changes were in deleted files, not really a conflict in utils.

Conflicts:
	src/replication/slave.cc
	src/replication/slave.hpp
	src/server/cmd_args.cc
	src/server/cmd_args.hpp
	src/server/key_value_store.cc
	src/server/server.cc
	src/utils.cc
	src/utils.hpp

+++ src/buffer_cache/mirrored/writeback.cc
@@ -361,7 +361,7 @@ void writeback_t::do_concurrent_flush() {
     ticks_t start_time2;
     pm_flushes_diff_flush.begin(&start_time2);
     // TODO: What are these magic constants doing here?
-    unsigned int blocks_to_flush = (unsigned long long)dirty_bufs.size() * 100ll / cache->get_block_size().value() + 1;
+    unsigned int blocks_to_flush = uint64_t(dirty_bufs.size()) * 100LL / cache->get_block_size().value() + 1;
     if (force_patch_storage_flush) {
         blocks_to_flush = std::max(cache->patch_disk_storage->get_number_of_log_blocks() / 20 + 1, blocks_to_flush);
         force_patch_storage_flush = false;
